Move ItemSource class to common library
